+++ src/ActivityLogStatsAssetBundle.php
@@ -20,7 +20,7 @@ class ActivityLogStatsAssetBundle extends AssetBundle
         // define the relative path to CSS/JS files that should be registered with the page
         // when this asset bundle is registered
         $this->js = [
-            'compiled/activity-log-stats.min.js?v=2',
+            'compiled/activity-log-stats.min.js?v=3',
         ];
 
         $this->css = [